+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- [swift] Add Swift syntax highlighting support
+
 ### Changed
 - License has been changed to LGPL 2.1 for the resulting hrc files with the specified license 'MPL 1.1/GPL 2.0/LGPL 2.1'.
 - Simplified catalog.xml.
@@ -55,6 +58,7 @@
 - [verilog] add define, include, timescale support
 - [cmake] fix escaping strings
 - [cpp] functions outliner list cleanup
+- [markdown] improve HTML handling (inline tags and CommonMark blocks)
 
 ### Added
 - New package type of base - all packed. Hrc and hrd files in one archive. Directory 'auto' not in archive.